finished refactoring app.py
