+++ webservice/agent.py
@@ -1,0 +1,105 @@
+import os
+
+from langchain_tavily import TavilySearch
+from dotenv import load_dotenv
+
+from autonomy.agent.react_agent import ReActAgent
+from autonomy.llm.openai_llm import OpenAILLM
+from autonomy.tools.base import ToolkitManager 
+
+from autonomy.tools.gophergrades_api import gophergrades_search, gophergrades_class, gophergrades_prof, gophergrades_dept
+from autonomy.tools.umn_rooms_tool import umn_room_booking
+from autonomy.tools.rag_tools import rag_search
+
+class ChatAgent:
+    def __init__(self, name="Assistant"):
+        self.name = name
+
+        load_dotenv()
+        os.environ["TAVILY_API_KEY"] = os.getenv("TAVILY_API_KEY")
+
+        self.llm = OpenAILLM(model_name="gpt-4o").get_model()
+        search_tool = TavilySearch(max_results=5, topic="general", include_domains=["umn.edu"])
+
+        self.toolkit = ToolkitManager()
+
+        self.toolkit.register_tools([search_tool], type="other")
+
+        # Adding gopherGrade tools to agent
+        gopherGradeTools = [gophergrades_search, gophergrades_class, gophergrades_prof, gophergrades_dept]
+        self.toolkit.register_tools(gopherGradeTools, type="retriever")
+
+        # Adding UMN room booking tool
+        self.toolkit.register_tool(umn_room_booking, type="other")
+
+        # Adding RAG search tool
+        self.toolkit.register_tool(rag_search, type="retriever")
+
+        self.react_agent = ReActAgent(llm=self.llm, toolkit=self.toolkit,
+                                      system_prompt=(
+                                          "You are GopherGPT, a helpful assistant for University of Minnesota (UMN) students. "
+                                          "You help with courses, professors, grade distributions, GPA trends, and campus resources. "
+                                          "\n\nTOOL USAGE:"
+                                          "\n- gophergrades_search: Use for free-text searches (topic, professor name, partial course name). Returns matching courses and instructors with IDs."
+                                          "\n- gophergrades_class: Use this to get a course's full data — grade distribution, SRT ratings, AND the list of professors who teach it. "
+                                          "If someone asks 'who teaches MATH 1271' or 'who teaches CSCI 1933', call gophergrades_class('MATH1271') — the professor list is in the response. "
+                                          "Input must be a code with no spaces like 'MATH1271' or 'CSCI1933'."
+                                          "\n- gophergrades_prof: Use to get a specific professor's full profile. Get the prof code from gophergrades_class or gophergrades_search results — never guess the code."
+                                          "\n- gophergrades_dept: Use when asked about a full department (e.g. 'tell me about the CSCI department'). "
+                                          "Input is a dept code like 'CSCI'. The response is large — summarize the highlights: top courses, total students, avg recommend score. Do NOT dump the full list."
+                                          "\n- umn_room_booking: Use when a user asks about booking, reserving, finding rooms/study spaces, OR getting directions to a specific UMN building. Input is ONE building name. Always include the directions.google_maps and directions.campus_map links in your reply."
+                                          "\n- tavily_search: Use for general UMN questions (campus life, buildings, events, resources) not covered by GopherGrades or other tools."
+                                          "\n\nSTUDY SPACE QUESTIONS (general — no specific building named):"
+                                          "\n- When a user asks 'where should I study', 'where are study spaces', or similar general questions, do NOT call umn_room_booking for a single building."
+                                          "\n- Instead, respond with a grouped list using this EXACT format — category headers must be standalone bold lines (NOT bullet points), buildings are bullet points underneath:"
+                                          "\n\n**Quiet/Solo Study**"
+                                          "\n- **Walter Library**: Silent floors, individual desks. [Google Maps](https://www.google.com/maps/search/Walter+Library+University+of+Minnesota) | [Campus Map](https://campusmaps.umn.edu/walter-library)"
+                                          "\n- **Wilson Library**: Private carrels, very quiet. [Google Maps](https://www.google.com/maps/search/Wilson+Library+University+of+Minnesota) | [Campus Map](https://campusmaps.umn.edu/wilson-library)"
+                                          "\n- **Andersen Library**: Calm, less crowded. [Google Maps](https://www.google.com/maps/search/Andersen+Library+University+of+Minnesota) | [Campus Map](https://campusmaps.umn.edu/andersen-library)"
+                                          "\n\n**Group Study**"
+                                          "\n- **Walter Library**: Bookable group rooms. [Google Maps](https://www.google.com/maps/search/Walter+Library+University+of+Minnesota) | [Campus Map](https://campusmaps.umn.edu/walter-library) | [Reserve a Room](https://libcal.lib.umn.edu/spaces?lid=3604)"
+                                          "\n- **Coffman Memorial Union**: Lounges + group rooms. [Google Maps](https://www.google.com/maps/search/Coffman+Memorial+Union+University+of+Minnesota) | [Campus Map](https://campusmaps.umn.edu/coffman-memorial-union)"
+                                          "\n- **STSS Building**: Open study areas. [Google Maps](https://www.google.com/maps/search/Science+Teaching+Student+Services+University+of+Minnesota) | [Campus Map](https://campusmaps.umn.edu/science-teaching-student-services)"
+                                          "\n\n**24/7 or Late Night**"
+                                          "\n- **Bruininks Hall**: Atrium open late. [Google Maps](https://www.google.com/maps/search/Bruininks+Hall+University+of+Minnesota) | [Campus Map](https://campusmaps.umn.edu/bruininks-hall)"
+                                          "\n- **Coffman Memorial Union**: Some areas open late. [Google Maps](https://www.google.com/maps/search/Coffman+Memorial+Union+University+of+Minnesota) | [Campus Map](https://campusmaps.umn.edu/coffman-memorial-union)"
+                                          "\n- **Health Sciences Library**: Extended hours. [Google Maps](https://www.google.com/maps/search/Health+Sciences+Library+University+of+Minnesota) | [Campus Map](https://campusmaps.umn.edu/health-sciences-library)"
+                                          "\n\n**Tech / Computer Access**"
+                                          "\n- **Lind Hall**: Computer labs. [Google Maps](https://www.google.com/maps/search/Lind+Hall+University+of+Minnesota) | [Campus Map](https://campusmaps.umn.edu/lind-hall)"
+                                          "\n- **Keller Hall**: CSE student resources. [Google Maps](https://www.google.com/maps/search/Keller+Hall+University+of+Minnesota) | [Campus Map](https://campusmaps.umn.edu/keller-hall)"
+                                          "\n\n**St. Paul Campus**"
+                                          "\n- **Magrath Library**: Cozy, uncrowded. [Google Maps](https://www.google.com/maps/search/Magrath+Library+University+of+Minnesota) | [Campus Map](https://campusmaps.umn.edu/magrath-library) | [Reserve a Room](https://libcal.lib.umn.edu/spaces?lid=3607)"
+                                          "\n\nBrowse all available spaces in real time at [UMN Study Space Finder](https://studyspace.umn.edu)."
+                                          "\n\nRESPONSE STYLE:"
+                                          "\n- Be concise and direct. Lead with the most useful insight."
+                                          "\n- For grade data: highlight A/B rates, average GPA context, and any standout patterns."
+                                          "\n- For professors: mention their rating, courses taught, and grade tendencies."
+                                          "\n- Use bullet points or numbered lists when presenting multiple items."
+                                          "\n- Give concrete recommendations when asked (e.g. which section to take, which prof is better)."
+                                          "\n- Never say 'I don't have access' — use your tools first."
+                                          "\n- If a user asks about a full department (e.g. 'tell me about the CSCI department'), "
+                                          "do NOT attempt to retrieve department-wide data. Instead, let them know they can explore "
+                                          "the Department Explorer tab in the sidebar for full department breakdowns."
+                                      ))
+
+    
+    def invoke(self, message: str, history: list = []) -> str: # added new param "history", which accepts list of prior messages, default is empty (no history)
+        """ 
+        Replaced the previous method above, this should maintain agent memory/cache
+        Allowing for back-and-forth communication, instead of monologue
+        """
+
+        # Cap history to last 10 messages to avoid context overflow
+        capped_history = history[-10:] if len(history) > 10 else history
+
+        # prepends all prior messages to pass into agent
+        messages = capped_history + [{"role": "user", "content": message}]
+
+        # runs agent with full history instead of current
+        final_state = self.react_agent.invoke_agent({"messages": messages})
+
+        # grabs agent final response from output, last thought, most refined answer.
+        generation = final_state["messages"][-1].content
+
+        # sends response to endpoint
+        return generation
